+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## 1.63.4
 
+### JavaScript API
+
+* Re-enable support for `import sass from 'sass'` when loading the package from
+  an ESM module in Node.js. However, this syntax is now deprecated; ESM users
+  should use `import * as sass from 'sass'` instead.
+
+  On the browser and other ESM-only platforms, only `import * as sass from
+  'sass'` is supported.
+
 ### Embedded Sass
 
 * Fix a race condition where closing standard input while requests are in-flight