+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 1.63.4
 
+### JavaScript API
+
+* Properly export the legacy API values `TRUE`, `FALSE`, `NULL`, and `types` from
+  the ECMAScript module API.
+
 ### Embedded Sass
 
 * Fix a race condition where closing standard input while requests are in-flight